cleanup approval modal
